requirements.txt
discord.py
discord.ext